requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
gunicorn==21.2.0
boto3==1.34.162
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.8.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0
